cleanup
- move test package definitions to testutil
- cleanup unused imports
- explicitly test each field instead of reimplementing the conversion
