+++ cosmos-server/src/test/scala/com/mesosphere/cosmos/converter/ResponseSpec.scala
@@ -5,7 +5,7 @@ import com.mesosphere.cosmos.converter.Common._
 import com.mesosphere.cosmos.thirdparty.marathon.model.AppId
 import com.mesosphere.cosmos.{label,ServiceMarathonTemplateNotFound, internal, rpc}
 import com.mesosphere.cosmos.converter.Response._
-import com.mesosphere.cosmos.PackageDefinitionTestObjects.{MinimalPackageDefinition,MaximalPackageDefinition}
+import com.mesosphere.cosmos.test.TestUtil.{MinimalPackageDefinition,MaximalPackageDefinition}
 
 import java.nio.charset.StandardCharsets
 
@@ -47,38 +47,36 @@ final class ResponseSpec extends FreeSpec {
 
       forAll(validV2s) { x => assertResult(Return(v1))(x.as[Try[rpc.v1.model.InstallResponse]]) }
     }
-
     "failure" in {
+      //expecting failure due to missing marathon mustache
       forAll(invalidV2s) { x => assertResult(Throw(ServiceMarathonTemplateNotFound(name, ver)))(x.as[Try[rpc.v1.model.InstallResponse]]) }
     }
   }
   "Conversion[internal.model.PackageDefinition,Try[rpc.v1.model.DescribeResponse]]" - {
     "success" in {
-      val d = universe.v2.model.PackageDetails(packagingVersion = MaximalPackageDefinition.packagingVersion.as[universe.v2.model.PackagingVersion],
-                                               name = MaximalPackageDefinition.name,
-                                               version = MaximalPackageDefinition.version.as[universe.v2.model.PackageDetailsVersion],
-                                               maintainer = MaximalPackageDefinition.maintainer,
-                                               description = MaximalPackageDefinition.description,
-                                               tags = MaximalPackageDefinition.tags.as[List[String]],
-                                               selected = Some(MaximalPackageDefinition.selected),
-                                               scm = MaximalPackageDefinition.scm,
-                                               website = MaximalPackageDefinition.website,
-                                               framework = Some(MaximalPackageDefinition.framework),
-                                               preInstallNotes = MaximalPackageDefinition.preInstallNotes,
-                                               postInstallNotes = MaximalPackageDefinition.postInstallNotes,
-                                               postUninstallNotes = MaximalPackageDefinition.postUninstallNotes,
-                                               licenses = MaximalPackageDefinition.licenses.as[Option[List[universe.v2.model.License]]])
-
       val m = MaximalPackageDefinition.marathon.get.v2AppMustacheTemplate
       val s = new String(ByteBuffers.getBytes(m), StandardCharsets.UTF_8)
-      val r = rpc.v1.model.DescribeResponse(`package` = d,
-                                            marathonMustache = s,
-                                            command = MaximalPackageDefinition.command.as[Option[universe.v2.model.Command]],
-                                            config = MaximalPackageDefinition.config,
-                                            resource = MaximalPackageDefinition.resource.map(_.as[universe.v2.model.Resource]))
-      assertResult(Return(r))(MaximalPackageDefinition.as[Try[rpc.v1.model.DescribeResponse]])
+      val conv = MaximalPackageDefinition.as[Try[rpc.v1.model.DescribeResponse]]
+      assert(conv.isReturn)
+      assertResult(s)                                                                                (conv.get.marathonMustache)
+      assertResult(MaximalPackageDefinition.command.as[Option[universe.v2.model.Command]])           (conv.get.command)                              
+      assertResult(MaximalPackageDefinition.config)                                                  (conv.get.config)                               
+      assertResult(MaximalPackageDefinition.resource.map(_.as[universe.v2.model.Resource]))          (conv.get.resource)                             
+      assertResult(MaximalPackageDefinition.packagingVersion.as[universe.v2.model.PackagingVersion]) (conv.get.`package`.packagingVersion)           
+      assertResult(MaximalPackageDefinition.name)                                                    (conv.get.`package`.name)                       
+      assertResult(MaximalPackageDefinition.version.as[universe.v2.model.PackageDetailsVersion])     (conv.get.`package`.version)                    
+      assertResult(MaximalPackageDefinition.maintainer)                                              (conv.get.`package`.maintainer)                 
+      assertResult(MaximalPackageDefinition.description)                                             (conv.get.`package`.description)                
+      assertResult(MaximalPackageDefinition.tags.as[List[String]])                                   (conv.get.`package`.tags)                       
+      assertResult(Some(MaximalPackageDefinition.selected))                                          (conv.get.`package`.selected)                   
+      assertResult(MaximalPackageDefinition.scm)                                                     (conv.get.`package`.scm)                        
+      assertResult(MaximalPackageDefinition.website)                                                 (conv.get.`package`.website)                    
+      assertResult(Some(MaximalPackageDefinition.framework))                                         (conv.get.`package`.framework)                  
+      assertResult(MaximalPackageDefinition.preInstallNotes)                                         (conv.get.`package`.preInstallNotes)            
+      assertResult(MaximalPackageDefinition.postInstallNotes)                                        (conv.get.`package`.postInstallNotes)           
+      assertResult(MaximalPackageDefinition.postUninstallNotes)                                      (conv.get.`package`.postUninstallNotes)         
+      assertResult(MaximalPackageDefinition.licenses.as[Option[List[universe.v2.model.License]]])    (conv.get.`package`.licenses)                   
     }
-
     "failure" in {
       //expecting failure due to missing marathon mustache
       val e = ServiceMarathonTemplateNotFound(MinimalPackageDefinition.name, MinimalPackageDefinition.version)
